Changed 'awesome conf' to 'awesome list of conf'
I'm increasingly unsure how many of these conferences are awesome as according to my guidelines, but I'm also hesitant to remove conferences without knowing if they are awesome or not. I figure that changing the name of the list would help - this is really just a list of conferences, not a list of great ones. For that, consult your own communities.