Overhaul list
The list needs an overhaul, badly. The conferences as they are aren't actually all the useful to anyone - some of these links are broken, some are old, and there's no general information on registration or CFPs. I've gone three and added a template and converted three as an example. I think that editing them and having a much fuller list would be good, for everyone.

Can others tell me if this looks like the right idea? Would you be keen on helping out?

Relevant to #3, #23, #34, #67, #69.