Overhaul the list to make it better (#94)
* Overhaul list

The list needs an overhaul, badly. The conferences as they are aren't actually all the useful to anyone - some of these links are broken, some are old, and there's no general information on registration or CFPs. I've gone three and added a template and converted three as an example. I think that editing them and having a much fuller list would be good, for everyone.

Can others tell me if this looks like the right idea? Would you be keen on helping out?

Relevant to #3, #23, #34, #67, #69.

* Move past conferences, add videos, notes, and roundups

* add DEVit conference (#98)

* Added npmCamp and new lines where appropriate 🎉 (#99)

Thought it was odd that there weren’t any new lines in the conference sections when there were new lines in every other section so I just added them. Also deleted last line.

I wanted to add npmCamp because I had such a great time as a diversity scholarship recipient. And since it’s marketed as a community conference focused on all things npm, I figured it deserved **awesome** recognition.

Signed-off-by: FVCproductions <hello@fvcproductions.com>

* add React Amsterdam and AmsterdamJS (#100)

* Update broken links (#101)

* adds active indgredients (#102)

* adds active indgredients

* Update README.md

* Add RigaDevDays to the list (#103)

* Add RigaDevDays to the list

* alphabetize

* add DevTernity conference (#104)

* Rename

See https://github.com/RichardLitt/awesome-conferences/issues/105

* Overhaul list

The list needs an overhaul, badly. The conferences as they are aren't actually all the useful to anyone - some of these links are broken, some are old, and there's no general information on registration or CFPs. I've gone three and added a template and converted three as an example. I think that editing them and having a much fuller list would be good, for everyone.

Can others tell me if this looks like the right idea? Would you be keen on helping out?

Relevant to #3, #23, #34, #67, #69.

* Move past conferences, add videos, notes, and roundups

* Overhaul

See #105